funciones para plathform math implementadas

+++ xcUtilities/include/xcVectorI4.h
@@ -18,150 +18,150 @@ class VectorI4 {
 	~VectorI4() = default;
 
 	/**
-		* @brief      + operator overload
-		* @param      V parameter one, vector for sum
-		* @bug	      No know Bugs
-		* @return     Returns sum of two vectors
-	*/
+	 * @brief      + operator overload
+	 * @param      V parameter one, vector for sum
+	 * @bug	       No know Bugs
+	 * @return     Returns sum of two vectors
+	 */
 	VectorI4
 	operator + (const VectorI4& V);
 
 	/**
-		* @brief      - operator overload
-		* @param      V parameter one, vector for rest
-		* @bug	      No know Bugs
-		* @return     Returns rest of two vectors
-	*/
+	 * @brief      - operator overload
+	 * @param      V parameter one, vector for rest
+	 * @bug	       No know Bugs
+	 * @return     Returns rest of two vectors
+	 */
 	VectorI4
 	operator - (const VectorI4& V);
 
 	/**
-		* @brief      * operator overload
-		* @param      V parameter one, vector for multiply
-		* @bug	      No know Bugs
-		* @return     Returns multiplication of two vectors
-	*/
+	 * @brief      * operator overload
+	 * @param      V parameter one, vector for multiply
+	 * @bug	       No know Bugs
+	 * @return     Returns multiplication of two vectors
+	 */
 	VectorI4
 	operator * (const VectorI4& V);
 
 	/**
-		* @brief      / operator overload
-		* @param      V parameter one, vector for divide
-		* @bug	      No know Bugs
-		* @return     Returns division of two vectors
-	*/
+	 * @brief      / operator overload
+	 * @param      V parameter one, vector for divide
+	 * @bug	       No know Bugs
+	 * @return     Returns division of two vectors
+	 */
 	VectorI4
 	operator / (const VectorI4& V);
 
 	/**
-		* @brief      + operator overload
-		* @param      V parameter one, vector for comparate
-		* @bug	      No know Bugs
-		* @return     Returns true if both vectors are equal
-	*/
+	 * @brief      + operator overload
+	 * @param      V parameter one, vector for comparative
+	 * @bug	       No know Bugs
+	 * @return     Returns true if both vectors are equal
+	 */
 	bool
 	operator == (const VectorI4& V);
 
 	/**
-		* @brief      + operator overload
-		* @param      V parameter one, number for sum
-		* @bug	      No know Bugs
-		* @return     Returns sum between a vector and a float
-	*/
+	 * @brief      + operator overload
+	 * @param      V parameter one, number for sum
+	 * @bug	       No know Bugs
+	 * @return     Returns sum between a vector and a float
+	 */
 	VectorI4
-	operator + (const float& V);
+	operator + (const int32& V);
 
 	/**
-		* @brief      + operator overload
-		* @param      V parameter one, number for rest
-		* @bug	      No know Bugs
-		* @return     Returns rest between a vector and a float
-	*/
+	 * @brief      + operator overload
+	 * @param      V parameter one, number for rest
+	 * @bug	       No know Bugs
+	 * @return     Returns rest between a vector and a float
+	 */
 	VectorI4
-	operator - (const float& V);
+	operator - (const int32& V);
 
 	/**
-		* @brief      + operator overload
-		* @param      V parameter one, number for multiply
-		* @bug	      No know Bugs
-		* @return     Returns multiplication between a vector and a float
-	*/
+	 * @brief      + operator overload
+	 * @param      V parameter one, number for multiply
+	 * @bug	       No know Bugs
+	 * @return     Returns multiplication between a vector and a float
+	 */
 	VectorI4
-	operator * (const float& V);
+	operator * (const int32& V);
 
 	/**
-		* @brief      + operator overload
-		* @param      V parameter one, number for divide
-		* @bug	      No know Bugs
-		* @return     Returns division between a vector and a float
-	*/
+	 * @brief      + operator overload
+	 * @param      V parameter one, number for divide
+	 * @bug	       No know Bugs
+	 * @return     Returns division between a vector and a float
+	 */
 	VectorI4
-	operator / (const float& V);
+	operator / (const int32& V);
 
 	/**
-		* @brief      += operator overload
-		* @param      V parameter one, vector for sum
-		* @bug	      No know Bugs
-		* @return     Returns the value of the sum of
-		*             your vector whit pother
-	*/
+	 * @brief      += operator overload
+	 * @param      V parameter one, vector for sum
+	 * @bug	       No know Bugs
+	 * @return     Returns the value of the sum of
+	 *             your vector whit pother
+	 */
 	VectorI4
 	operator += (const VectorI4& V);
 
 	/**
-		* @brief      -= operator overload
-		* @param      V parameter one, vector for sum
-		* @bug	      No know Bugs
-		* @return     Returns the value of the rest of
-		*             your vector whit pother
-	*/
+	 * @brief      -= operator overload
+	 * @param      V parameter one, vector for sum
+	 * @bug	       No know Bugs
+	 * @return     Returns the value of the rest of
+	 *             your vector whit pother
+	 */
 	VectorI4
 	operator -= (const VectorI4& V);
 
 	/**
-		* @brief      *= operator overload
-		* @param      V parameter one, vector for sum
-		* @bug	      No know Bugs
-		* @return     Returns the value of the multiplication of
-		*             your vector whit pother
-	*/
+	 * @brief      *= operator overload
+	 * @param      V parameter one, vector for sum
+	 * @bug	       No know Bugs
+	 * @return     Returns the value of the multiplication of
+	 *             your vector whit pother
+	 */
 	VectorI4
 	operator *= (const VectorI4& V);
 
 	/**
-		* @brief      /= operator overload
-		* @param      V parameter one, vector for sum
-		* @bug	      No know Bugs
-		* @return     Returns the value of the division of
-		*             your vector whit pother
-	*/
+	 * @brief      /= operator overload
+	 * @param      V parameter one, vector for sum
+	 * @bug	       No know Bugs
+	 * @return     Returns the value of the division of
+	 *             your vector whit pother
+	 */
 	VectorI4
 	operator /= (const VectorI4& V);
 
 
  public:
 	/**
-		* @brief Public variable member
-	*/
+	 * @brief Public variable member
+	 */
 
 	/**
-		* @Variable m_x
-	*/
+	 * @Variable m_x
+	 */
 	int32 m_x;
 
 	/**
-		* @Variable m_y
-	*/
+	 * @Variable m_y
+	 */
 	int32 m_y;
 
 	/**
-		* @Variable m_z
-	*/
+	 * @Variable m_z
+	 */
 	int32 m_z;
 
 	/**
-		* @Variable m_z
-	*/
+	 * @Variable m_z
+	 */
 	int32 m_w;
 
 };
